chore: rename `compiler.rs` to `ecma_compiler.rs`
