+++ crates/rolldown_ecmascript/src/ecma_compiler.rs
