+++ crates/rolldown_ecmascript/src/lib.rs
@@ -1,14 +1,14 @@
 mod allocator_helpers;
 mod ast_snippet;
-mod compiler;
 mod ecma_ast;
+mod ecma_compiler;
 mod ext;
 
 pub use crate::{
   allocator_helpers::take_in::TakeIn,
   ast_snippet::AstSnippet,
-  compiler::EcmaCompiler,
   ecma_ast::{program_cell::WithMutFields, EcmaAst},
+  ecma_compiler::EcmaCompiler,
   ext::{
     allocator_ext::AllocatorExt, span_ext::SpanExt, BindingIdentifierExt, BindingPatternExt,
     ExpressionExt, StatementExt,
